Fixing BiggyList FlushToDisk method
Changing the serialize object to a string builder and serialize each
object individually as the string replace ("[", "]", ",") causes
problems when a string property has the characters in.

Also changed from using the file writer to File.WriteAllText as the
filewriter caused problems when the new json length was less than the
old length.